fix: replace HashSet with ConcurrentDictionary for thread-safe ArrayPool tracking
The ArrayPool tracking system was using a non-thread-safe HashSet, causing race conditions and test failures under concurrent access. Replaced with ConcurrentDictionary to ensure thread safety.

Also fixed bounds checking bug in StressTest_ConcurrentShapeDeduplication where array index could exceed shape height.
