+++ Assets/Tests/DopeGrid/ImmutableGridShapeConcurrencyTests.cs
@@ -1,21 +1,18 @@
-#if UNITY_5_3_OR_NEWER
 using System;
 using System.Collections.Generic;
-using System.ComponentModel;
-using System.Diagnostics;
 using System.Linq;
 using System.Threading;
+using System.Threading.Tasks;
 using DopeGrid;
-using DopeGrid.Native;
 using NUnit.Framework;
-using Unity.Collections;
-using Unity.Jobs;
-using Unity.Mathematics;
-using UnityEngine;
-using Random = Unity.Mathematics.Random;
 
+namespace DopeGrid.Tests;
+
+/// <summary>
+/// Standard C# concurrency tests for ImmutableGridShape (no Unity dependencies)
+/// </summary>
 [TestFixture]
-public class ImmutableGridShapeConcurrencyTests
+public class ImmutableGridShapeConcurrencyStandardTests
 {
     [TearDown]
     public void Cleanup()
@@ -43,10 +40,9 @@ public class ImmutableGridShapeConcurrencyTests
             {
                 barrier.SignalAndWait(); // Ensure all threads start simultaneously
 
-                var shape = Shapes.TShape(Allocator.TempJob);
-                var immutable = shape.AsReadOnly().GetOrCreateImmutable();
+                using var shape = Shapes.TShape();
+                var immutable = shape.GetOrCreateImmutable();
                 results[index] = immutable.Id;
-                shape.Dispose();
             });
         }
 
@@ -55,7 +51,7 @@ public class ImmutableGridShapeConcurrencyTests
 
         // All threads should get the same ID for the same shape
         var firstId = results[0];
-        Assert.IsTrue(results.All(id => id == firstId),
+        Assert.That(results.All(id => id == firstId), Is.True,
             $"Expected all IDs to be {firstId}, but got: [{string.Join(", ", results)}]");
     }
 
@@ -75,15 +71,9 @@ public class ImmutableGridShapeConcurrencyTests
                 barrier.SignalAndWait(); // Ensure all threads start simultaneously
 
                 // Each thread creates a unique shape
-                var shape = new GridShape(index + 1, 1, Allocator.TempJob);
-                for (int j = 0; j <= index; j++)
-                {
-                    shape.SetCell(new int2(j, 0), true);
-                }
-
-                var immutable = shape.AsReadOnly().GetOrCreateImmutable();
+                using var shape = Shapes.Line(index + 1);
+                var immutable = shape.GetOrCreateImmutable();
                 results[index] = immutable.Id;
-                shape.Dispose();
             });
         }
 
@@ -92,7 +82,7 @@ public class ImmutableGridShapeConcurrencyTests
 
         // All threads should get different IDs
         var uniqueIds = new HashSet<int>(results);
-        Assert.AreEqual(threadCount, uniqueIds.Count,
+        Assert.That(uniqueIds.Count, Is.EqualTo(threadCount),
             $"Expected {threadCount} unique IDs, but got: [{string.Join(", ", results)}]");
     }
 
@@ -100,9 +90,8 @@ public class ImmutableGridShapeConcurrencyTests
     public void ConcurrentRotations_SameShape_ConsistentResults()
     {
         // First, create a base L-shape
-        var baseShape = Shapes.LShape(Allocator.TempJob);
-        var baseImmutable = baseShape.AsReadOnly().GetOrCreateImmutable();
-        baseShape.Dispose();
+        using var baseShape = Shapes.LShape();
+        var baseImmutable = baseShape.GetOrCreateImmutable();
 
         const int threadCount = 20;
         var results = new (int rot90, int rot180, int rot270)[threadCount];
@@ -116,9 +105,9 @@ public class ImmutableGridShapeConcurrencyTests
             {
                 barrier.SignalAndWait(); // Ensure all threads start simultaneously
 
-                var rot90 = baseImmutable.Rotate90(Allocator.TempJob);
-                var rot180 = rot90.Rotate90(Allocator.TempJob);
-                var rot270 = rot180.Rotate90(Allocator.TempJob);
+                var rot90 = baseImmutable.Rotate90();
+                var rot180 = rot90.Rotate90();
+                var rot270 = rot180.Rotate90();
 
                 results[index] = (rot90.Id, rot180.Id, rot270.Id);
             });
@@ -129,11 +118,11 @@ public class ImmutableGridShapeConcurrencyTests
 
         // All threads should get the same rotation IDs
         var first = results[0];
-        Assert.IsTrue(results.All(r => r.rot90 == first.rot90),
+        Assert.That(results.All(r => r.rot90 == first.rot90), Is.True,
             "Inconsistent 90-degree rotation IDs");
-        Assert.IsTrue(results.All(r => r.rot180 == first.rot180),
+        Assert.That(results.All(r => r.rot180 == first.rot180), Is.True,
             "Inconsistent 180-degree rotation IDs");
-        Assert.IsTrue(results.All(r => r.rot270 == first.rot270),
+        Assert.That(results.All(r => r.rot270 == first.rot270), Is.True,
             "Inconsistent 270-degree rotation IDs");
     }
 
@@ -141,9 +130,8 @@ public class ImmutableGridShapeConcurrencyTests
     public void ConcurrentFlips_SameShape_ConsistentResults()
     {
         // Create a base shape that's asymmetric
-        var baseShape = Shapes.LShape(Allocator.TempJob);
-        var baseImmutable = baseShape.AsReadOnly().GetOrCreateImmutable();
-        baseShape.Dispose();
+        using var baseShape = Shapes.LShape();
+        var baseImmutable = baseShape.GetOrCreateImmutable();
 
         const int threadCount = 20;
         var results = new int[threadCount];
@@ -157,7 +145,7 @@ public class ImmutableGridShapeConcurrencyTests
             {
                 barrier.SignalAndWait(); // Ensure all threads start simultaneously
 
-                var flipped = baseImmutable.Flip(Allocator.TempJob);
+                var flipped = baseImmutable.Flip();
                 results[index] = flipped.Id;
             });
         }
@@ -167,7 +155,7 @@ public class ImmutableGridShapeConcurrencyTests
 
         // All threads should get the same flipped ID
         var firstId = results[0];
-        Assert.IsTrue(results.All(id => id == firstId),
+        Assert.That(results.All(id => id == firstId), Is.True,
             $"Expected all flipped IDs to be {firstId}, but got: [{string.Join(", ", results)}]");
     }
 
@@ -189,52 +177,51 @@ public class ImmutableGridShapeConcurrencyTests
                 {
                     barrier.SignalAndWait();
 
-                    var random = new Random((uint)(threadId + 1));
+                    var random = new Random(threadId + 1);
                     var createdShapes = new List<ImmutableGridShape>();
 
                     for (int op = 0; op < operationsPerThread; op++)
                     {
-                        var operation = random.NextInt(0, 4);
+                        var operation = random.Next(0, 4);
 
                         switch (operation)
                         {
                             case 0: // Create new shape
                             {
-                                var width = random.NextInt(1, 5);
-                                var height = random.NextInt(1, 5);
-                                var shape = new GridShape(width, height, Allocator.TempJob);
+                                var width = random.Next(1, 5);
+                                var height = random.Next(1, 5);
+                                using var shape = new GridShape(width, height);
 
                                 // Set random cells
                                 for (int x = 0; x < width; x++)
                                 {
                                     for (int y = 0; y < height; y++)
                                     {
-                                        if (random.NextBool())
-                                            shape.SetCell(new int2(x, y), true);
+                                        if (random.Next(0, 2) == 1)
+                                            shape[x, y] = true;
                                     }
                                 }
 
                                 // Ensure at least one cell is set
-                                shape.SetCell(new int2(0, 0), true);
+                                shape[0, 0] = true;
 
-                                var trimmed = shape.AsReadOnly().Trim(Allocator.TempJob);
-                                var immutable = trimmed.AsReadOnly().GetOrCreateImmutable();
+                                var (_, _, trimWidth, trimHeight) = shape.GetTrimmedBound<GridShape, bool>();
+                                using var trimmed = new GridShape(trimWidth, trimHeight);
+                                shape.Trim<GridShape, GridShape, bool>(trimmed);
+                                var immutable = trimmed.GetOrCreateImmutable();
                                 createdShapes.Add(immutable);
-
-                                trimmed.Dispose();
-                                shape.Dispose();
                                 break;
                             }
                             case 1: // Rotate existing shape
                             {
                                 if (createdShapes.Count > 0)
                                 {
-                                    var shape = createdShapes[random.NextInt(0, createdShapes.Count)];
-                                    var rotated = shape.Rotate90(Allocator.TempJob);
+                                    var shape = createdShapes[random.Next(0, createdShapes.Count)];
+                                    var rotated = shape.Rotate90();
 
                                     // Verify rotation properties
-                                    var rotated360 = rotated.Rotate90(Allocator.TempJob).Rotate90(Allocator.TempJob).Rotate90(Allocator.TempJob);
-                                    Assert.AreEqual(shape.Id, rotated360.Id,
+                                    var rotated360 = rotated.Rotate90().Rotate90().Rotate90();
+                                    Assert.That(rotated360.Id, Is.EqualTo(shape.Id),
                                         "360-degree rotation should return to original");
                                 }
                                 break;
@@ -243,12 +230,12 @@ public class ImmutableGridShapeConcurrencyTests
                             {
                                 if (createdShapes.Count > 0)
                                 {
-                                    var shape = createdShapes[random.NextInt(0, createdShapes.Count)];
-                                    var flipped = shape.Flip(Allocator.TempJob);
+                                    var shape = createdShapes[random.Next(0, createdShapes.Count)];
+                                    var flipped = shape.Flip();
 
                                     // Verify flip properties
-                                    var doubleFlipped = flipped.Flip(Allocator.TempJob);
-                                    Assert.AreEqual(shape.Id, doubleFlipped.Id,
+                                    var doubleFlipped = flipped.Flip();
+                                    Assert.That(doubleFlipped.Id, Is.EqualTo(shape.Id),
                                         "Double flip should return to original");
                                 }
                                 break;
@@ -257,14 +244,14 @@ public class ImmutableGridShapeConcurrencyTests
                             {
                                 if (createdShapes.Count > 0)
                                 {
-                                    var shape = createdShapes[random.NextInt(0, createdShapes.Count)];
+                                    var shape = createdShapes[random.Next(0, createdShapes.Count)];
 
                                     // Verify basic properties are consistent
-                                    Assert.GreaterOrEqual(shape.Width, 0);
-                                    Assert.GreaterOrEqual(shape.Height, 0);
-                                    Assert.AreEqual(shape.Width * shape.Height, shape.Size);
-                                    Assert.GreaterOrEqual(shape.OccupiedSpaceCount, 0);
-                                    Assert.LessOrEqual(shape.OccupiedSpaceCount, shape.Size);
+                                    Assert.That(shape.Width, Is.GreaterThanOrEqualTo(0));
+                                    Assert.That(shape.Height, Is.GreaterThanOrEqualTo(0));
+                                    Assert.That(shape.Size, Is.EqualTo(shape.Width * shape.Height));
+                                    Assert.That(shape.OccupiedSpaceCount(), Is.GreaterThanOrEqualTo(0));
+                                    Assert.That(shape.OccupiedSpaceCount(), Is.LessThanOrEqualTo(shape.Size));
                                 }
                                 break;
                             }
@@ -284,331 +271,163 @@ public class ImmutableGridShapeConcurrencyTests
         foreach (var thread in threads) thread.Start();
         foreach (var thread in threads) thread.Join();
 
-        Assert.IsEmpty(errors, $"Errors occurred: {string.Join("\n", errors)}");
+        Assert.That(errors, Is.Empty, $"Errors occurred: {string.Join("\n", errors)}");
     }
 
     #endregion
 
-#if UNITY_2022_3_OR_NEWER
-    #region Unity Job System Tests
+    #region Task-based Async Tests
 
-    struct CreateShapeJob : IJobParallelFor
+    [Test]
+    public async Task AsyncConcurrentCreation_SameShape_ReturnsSameId()
     {
-        public int ShapeSize;
-        public NativeArray<int> Results;
+        const int taskCount = 20;
+        var tasks = new Task<int>[taskCount];
 
-        public void Execute(int index)
+        for (int i = 0; i < taskCount; i++)
         {
-            var shape = new GridShape(ShapeSize, ShapeSize, Allocator.Temp);
-
-            // Create a unique pattern based on index
-            for (int i = 0; i <= index && i < ShapeSize * ShapeSize; i++)
+            tasks[i] = Task.Run(() =>
             {
-                shape.SetCell(new int2(i % ShapeSize, i / ShapeSize), true);
-            }
-
-            var trimmed = shape.AsReadOnly().Trim(Allocator.Temp);
-            var immutable = trimmed.AsReadOnly().GetOrCreateImmutable();
-            Results[index] = immutable.Id;
-
-            trimmed.Dispose();
-            shape.Dispose();
+                using var shape = Shapes.TShape();
+                return shape.GetOrCreateImmutable().Id;
+            });
         }
+
+        var results = await Task.WhenAll(tasks).ConfigureAwait(false);
+
+        // All tasks should get the same ID
+        var firstId = results[0];
+        Assert.That(results.All(id => id == firstId), Is.True,
+            $"Expected all IDs to be {firstId}, but got: [{string.Join(", ", results)}]");
     }
 
     [Test]
-    public void JobSystem_ParallelShapeCreation_ConsistentIds()
+    public async Task AsyncConcurrentTransformations_ConsistentResults()
     {
-        const int jobCount = 10;
-        var results = new NativeArray<int>(jobCount, Allocator.TempJob);
+        using var baseShape = Shapes.LShape();
+        var baseImmutable = baseShape.GetOrCreateImmutable();
 
-        try
+        const int taskCount = 30;
+        var tasks = new Task<(int rotId, int flipId)>[taskCount];
+
+        for (int i = 0; i < taskCount; i++)
         {
-            var job = new CreateShapeJob
+            tasks[i] = Task.Run(() =>
             {
-                ShapeSize = 3,
-                Results = results
-            };
-
-            var handle = job.Schedule(jobCount, 1);
-            handle.Complete();
-
-            // After trimming, shapes with the same pattern should get the same ID
-            // Different patterns should get different IDs
-            var idToIndices = new Dictionary<int, List<int>>();
-            for (int i = 0; i < jobCount; i++)
-            {
-                if (!idToIndices.ContainsKey(results[i]))
-                {
-                    idToIndices[results[i]] = new List<int>();
-                }
-                idToIndices[results[i]].Add(i);
-            }
-
-            // Verify that all results have valid IDs (greater than 0)
-            for (int i = 0; i < jobCount; i++)
-            {
-                Assert.Greater(results[i], 0, $"Job {i} returned invalid ID");
-            }
-
-            // Log the groupings for debugging
-            Debug.Log($"Shape creation resulted in {idToIndices.Count} unique shapes from {jobCount} jobs");
-            foreach (var kvp in idToIndices)
-            {
-                Debug.Log($"Shape ID {kvp.Key} was created by jobs: [{string.Join(", ", kvp.Value)}]");
-            }
+                var rotated = baseImmutable.Rotate90();
+                var flipped = baseImmutable.Flip();
+                return (rotated.Id, flipped.Id);
+            });
         }
-        finally
-        {
-            results.Dispose();
-        }
-    }
 
-    struct RotateShapeJob : IJob
-    {
-        public int BaseShapeId;
-        public NativeArray<int> RotationResults;
+        var results = await Task.WhenAll(tasks).ConfigureAwait(false);
 
-        public void Execute()
-        {
-            var baseShape = new ImmutableGridShape(BaseShapeId);
-
-            var rot90 = baseShape.Rotate90(Allocator.Temp);
-            var rot180 = rot90.Rotate90(Allocator.Temp);
-            var rot270 = rot180.Rotate90(Allocator.Temp);
-            var rot360 = rot270.Rotate90(Allocator.Temp);
-
-            RotationResults[0] = rot90.Id;
-            RotationResults[1] = rot180.Id;
-            RotationResults[2] = rot270.Id;
-            RotationResults[3] = rot360.Id;
-        }
+        // All tasks should get the same transformation IDs
+        var first = results[0];
+        Assert.That(results.All(r => r.rotId == first.rotId), Is.True,
+            "Inconsistent rotation IDs");
+        Assert.That(results.All(r => r.flipId == first.flipId), Is.True,
+            "Inconsistent flip IDs");
     }
 
     [Test]
-    public void JobSystem_MultipleRotationJobs_ConsistentResults()
+    public async Task AsyncStressTest_ParallelOperations()
     {
-        // Create a base shape first
-        var baseShape = Shapes.LShape(Allocator.TempJob);
-        var baseImmutable = baseShape.AsReadOnly().GetOrCreateImmutable();
-        baseShape.Dispose();
+        const int taskCount = 50;
+        const int operationsPerTask = 100;
+        var errors = new List<string>();
+        var errorsLock = new object();
 
-        const int jobCount = 10;
-        var jobs = new NativeArray<JobHandle>(jobCount, Allocator.TempJob);
-        var results = new NativeArray<int>[jobCount];
-
-        try
+        var tasks = Enumerable.Range(0, taskCount).Select(taskId => Task.Run(() =>
         {
-            // Schedule multiple rotation jobs
-            for (int i = 0; i < jobCount; i++)
+            try
             {
-                results[i] = new NativeArray<int>(4, Allocator.TempJob);
-                var job = new RotateShapeJob
+                var random = new Random(taskId);
+                var successCount = 0;
+
+                for (int op = 0; op < operationsPerTask; op++)
                 {
-                    BaseShapeId = baseImmutable.Id,
-                    RotationResults = results[i]
-                };
-                jobs[i] = job.Schedule();
-            }
+                    var operation = random.Next(0, 3);
 
-            // Wait for all jobs to complete
-            JobHandle.CompleteAll(jobs);
-
-            // Verify all jobs got the same rotation results
-            var firstResult = results[0];
-            for (int i = 1; i < jobCount; i++)
-            {
-                Assert.AreEqual(firstResult[0], results[i][0], $"Job {i} got different 90-degree rotation");
-                Assert.AreEqual(firstResult[1], results[i][1], $"Job {i} got different 180-degree rotation");
-                Assert.AreEqual(firstResult[2], results[i][2], $"Job {i} got different 270-degree rotation");
-                Assert.AreEqual(baseImmutable.Id, results[i][3], $"Job {i} 360-degree rotation didn't return to original");
-            }
-        }
-        finally
-        {
-            jobs.Dispose();
-            foreach (var result in results)
-            {
-                if (result.IsCreated)
-                    result.Dispose();
-            }
-        }
-    }
-
-    struct StressTestJob : IJobParallelFor
-    {
-        public int OperationsPerJob;
-        [ReadOnly] public NativeArray<Random> RandomGenerators;
-        public NativeArray<int> SuccessCount;
-        public NativeArray<int> FailureCount;
-        public NativeArray<int> FirstFailureOperation;
-
-        public void Execute(int index)
-        {
-            var random = RandomGenerators[index];
-            var successfulOps = 0;
-            var failedOps = 0;
-            var firstFailure = -1;
-
-            for (int op = 0; op < OperationsPerJob; op++)
-            {
-                var operation = random.NextInt(0, 3);
-
-                try
-                {
                     switch (operation)
                     {
                         case 0: // Create random shape
                         {
-                            var size = random.NextInt(1, 4);
-                            var shape = new GridShape(size, size, Allocator.Temp);
-                            shape.SetCell(new int2(0, 0), true);
+                            var size = random.Next(1, 4);
+                            using var shape = new GridShape(size, size);
+                            shape[0, 0] = true;
 
-                            if (random.NextBool())
-                                shape.SetCell(new int2(size - 1, size - 1), true);
+                            if (random.Next(0, 2) == 1)
+                                shape[size - 1, size - 1] = true;
 
-                            var trimmed = shape.AsReadOnly().Trim(Allocator.Temp);
-                            var immutable = trimmed.AsReadOnly().GetOrCreateImmutable();
+                            var (_, _, trimWidth, trimHeight) = shape.GetTrimmedBound<GridShape, bool>();
+                            using var trimmed = new GridShape(trimWidth, trimHeight);
+                            shape.Trim<GridShape, GridShape, bool>(trimmed);
+                            var immutable = trimmed.GetOrCreateImmutable();
 
                             if (immutable.Id > 0)
-                                successfulOps++;
-                            else
-                                throw new Exception($"Invalid ID returned: {immutable.Id}");
-
-                            trimmed.Dispose();
-                            shape.Dispose();
+                                successCount++;
                             break;
                         }
                         case 1: // Test standard shapes
                         {
-                            var shapeType = random.NextInt(0, 4);
-                            GridShape testShape;
-
-                            switch (shapeType)
+                            var shapeType = random.Next(0, 4);
+                            using GridShape testShape = shapeType switch
                             {
-                                case 0: testShape = Shapes.Line(3, Allocator.Temp); break;
-                                case 1: testShape = Shapes.Square(2, Allocator.Temp); break;
-                                case 2: testShape = Shapes.LShape(Allocator.Temp); break;
-                                default: testShape = Shapes.TShape(Allocator.Temp); break;
-                            }
+                                0 => Shapes.Line(3),
+                                1 => Shapes.Square(2),
+                                2 => Shapes.LShape(),
+                                _ => Shapes.TShape()
+                            };
 
-                            var immutable = testShape.AsReadOnly().GetOrCreateImmutable();
-                            var rotated = immutable.Rotate90(Allocator.Temp);
-                            var flipped = immutable.Flip(Allocator.Temp);
+                            var immutable = testShape.GetOrCreateImmutable();
+                            var rotated = immutable.Rotate90();
+                            var flipped = immutable.Flip();
 
                             if (rotated.Id > 0 && flipped.Id > 0)
-                                successfulOps++;
-                            else
-                                throw new Exception($"Invalid transform IDs: rot={rotated.Id}, flip={flipped.Id}");
-
-                            testShape.Dispose();
+                                successCount++;
                             break;
                         }
                         case 2: // Verify shape properties
                         {
-                            var shape = Shapes.Cross(Allocator.Temp);
-                            var immutable = shape.AsReadOnly().GetOrCreateImmutable();
+                            using var shape = Shapes.Cross();
+                            var immutable = shape.GetOrCreateImmutable();
 
                             if (immutable.Width > 0 && immutable.Height > 0 &&
-                                immutable.OccupiedSpaceCount > 0)
+                                immutable.OccupiedSpaceCount() > 0)
                             {
-                                successfulOps++;
+                                successCount++;
                             }
-                            else
-                            {
-                                throw new Exception($"Invalid properties: W={immutable.Width}, H={immutable.Height}, Occ={immutable.OccupiedSpaceCount}");
-                            }
-
-                            shape.Dispose();
                             break;
                         }
                     }
                 }
-                catch (Exception e)
+
+                return successCount;
+            }
+            catch (Exception ex)
+            {
+                lock (errorsLock)
                 {
-                    failedOps++;
-                    if (firstFailure == -1)
-                    {
-                        firstFailure = op * 10 + operation; // Encode both op index and operation type
-                        Debug.LogError($"Job {index}, Op {op}, Type {operation} failed: {e.Message}");
-                    }
+                    errors.Add($"Task {taskId}: {ex.Message}");
                 }
+                return 0;
             }
+        })).ToArray();
 
-            SuccessCount[index] = successfulOps;
-            FailureCount[index] = failedOps;
-            FirstFailureOperation[index] = firstFailure;
-        }
-    }
+        var results = await Task.WhenAll(tasks).ConfigureAwait(false);
 
-    [Test]
-    public void JobSystem_StressTest_HighConcurrency()
-    {
-        const int jobCount = 50;
-        const int operationsPerJob = 100;
+        Assert.That(errors, Is.Empty, $"Errors occurred:\n{string.Join("\n", errors)}");
 
-        var randomGenerators = new NativeArray<Random>(jobCount, Allocator.TempJob);
-        var successCounts = new NativeArray<int>(jobCount, Allocator.TempJob);
-        var failureCounts = new NativeArray<int>(jobCount, Allocator.TempJob);
-        var firstFailureOps = new NativeArray<int>(jobCount, Allocator.TempJob);
+        var totalSuccess = results.Sum();
+        var expectedMinSuccess = taskCount * operationsPerTask * 0.9; // Expect at least 90% success
 
-        try
-        {
-            // Initialize random generators outside the job
-            var systemRandom = new System.Random(42);
-            for (int i = 0; i < jobCount; i++)
-            {
-                randomGenerators[i] = new Random((uint)systemRandom.Next());
-            }
-
-            var job = new StressTestJob
-            {
-                OperationsPerJob = operationsPerJob,
-                RandomGenerators = randomGenerators,
-                SuccessCount = successCounts,
-                FailureCount = failureCounts,
-                FirstFailureOperation = firstFailureOps
-            };
-
-            var handle = job.Schedule(jobCount, 1);
-            handle.Complete();
-
-            // Verify all jobs completed successfully
-            var totalSuccess = 0;
-            var totalFailures = 0;
-            for (int i = 0; i < jobCount; i++)
-            {
-                totalSuccess += successCounts[i];
-                totalFailures += failureCounts[i];
-
-                if (failureCounts[i] > 0)
-                {
-                    var firstFail = firstFailureOps[i];
-                    var opIndex = firstFail / 10;
-                    var opType = firstFail % 10;
-                    Debug.LogWarning($"Job {i}: {successCounts[i]} successes, {failureCounts[i]} failures. First failure at op {opIndex}, type {opType}");
-                }
-
-                Assert.Greater(successCounts[i], 0, $"Job {i} had no successful operations");
-            }
-
-            Debug.Log($"Total: {totalSuccess} successes, {totalFailures} failures out of {jobCount * operationsPerJob} operations");
-
-            var expectedMinSuccess = jobCount * operationsPerJob * 0.8f; // Expect at least 80% success
-            Assert.Greater(totalSuccess, expectedMinSuccess,
-                $"Total successful operations ({totalSuccess}) below expected minimum ({expectedMinSuccess})");
-        }
-        finally
-        {
-            randomGenerators.Dispose();
-            successCounts.Dispose();
-            failureCounts.Dispose();
-            firstFailureOps.Dispose();
-        }
+        Assert.That(totalSuccess, Is.GreaterThan(expectedMinSuccess),
+            $"Total successful operations ({totalSuccess}) below expected minimum ({expectedMinSuccess})");
     }
 
     #endregion
-#endif
+
     #region Stress Tests
 
     [Test]
@@ -629,28 +448,30 @@ public class ImmutableGridShapeConcurrencyTests
             {
                 startSignal.Wait(); // Wait for all threads to be ready
 
-                var random = new Random((uint)(threadIndex + 1));
+                var random = new Random(threadIndex + 1);
 
                 for (int s = 0; s < shapesPerThread; s++)
                 {
                     // Create shapes with varying complexity
-                    var complexity = random.NextInt(1, 5);
-                    var shape = new GridShape(complexity, complexity, Allocator.TempJob);
+                    var complexity = random.Next(1, 5);
+                    using var shape = new GridShape(complexity, complexity);
 
                     // Random pattern
                     for (int i = 0; i < complexity * complexity; i++)
                     {
-                        if (random.NextFloat() > 0.5f)
+                        if (random.NextDouble() > 0.5)
                         {
-                            shape.SetCell(new int2(i % complexity, i / complexity), true);
+                            shape[i % complexity, i / complexity] = true;
                         }
                     }
 
                     // Ensure at least one cell
-                    shape.SetCell(new int2(0, 0), true);
+                    shape[0, 0] = true;
 
-                    var trimmed = shape.AsReadOnly().Trim(Allocator.TempJob);
-                    var immutable = trimmed.AsReadOnly().GetOrCreateImmutable();
+                    var (_, _, trimWidth, trimHeight) = shape.GetTrimmedBound<GridShape, bool>();
+                    using var trimmed = new GridShape(trimWidth, trimHeight);
+                    shape.Trim<GridShape, GridShape, bool>(trimmed);
+                    var immutable = trimmed.GetOrCreateImmutable();
 
                     lock (allIds[threadIndex])
                     {
@@ -658,26 +479,23 @@ public class ImmutableGridShapeConcurrencyTests
                     }
 
                     // Also test transformations
-                    if (random.NextBool())
+                    if (random.Next(0, 2) == 1)
                     {
-                        var rotated = immutable.Rotate90(Allocator.TempJob);
+                        var rotated = immutable.Rotate90();
                         lock (allIds[threadIndex])
                         {
                             allIds[threadIndex].Add(rotated.Id);
                         }
                     }
 
-                    if (random.NextBool())
+                    if (random.Next(0, 2) == 1)
                     {
-                        var flipped = immutable.Flip(Allocator.TempJob);
+                        var flipped = immutable.Flip();
                         lock (allIds[threadIndex])
                         {
                             allIds[threadIndex].Add(flipped.Id);
                         }
                     }
-
-                    trimmed.Dispose();
-                    shape.Dispose();
                 }
             });
         }
@@ -692,27 +510,27 @@ public class ImmutableGridShapeConcurrencyTests
 
         // Verify results
         var totalIds = allIds.SelectMany(list => list).ToList();
-        Assert.Greater(totalIds.Count, 0, "No shapes were created");
-        Assert.IsTrue(totalIds.All(id => id > 0), "Some IDs were invalid (<=0)");
+        Assert.That(totalIds.Count, Is.GreaterThan(0), "No shapes were created");
+        Assert.That(totalIds.All(id => id > 0), Is.True, "Some IDs were invalid (<=0)");
 
-        Debug.Log($"Stress test created {totalIds.Distinct().Count()} unique shapes across {threadCount} threads");
+        TestContext.WriteLine($"Stress test created {totalIds.Distinct().Count()} unique shapes across {threadCount} threads");
     }
 
     [Test]
     public void StressTest_RapidTransformations()
     {
         // Create a complex base shape
-        var shape = new GridShape(4, 4, Allocator.TempJob);
-        shape.SetCell(new int2(0, 0), true);
-        shape.SetCell(new int2(1, 0), true);
-        shape.SetCell(new int2(0, 1), true);
-        shape.SetCell(new int2(2, 2), true);
-        shape.SetCell(new int2(3, 3), true);
+        using var shape = new GridShape(4, 4);
+        shape[0, 0] = true;
+        shape[1, 0] = true;
+        shape[0, 1] = true;
+        shape[2, 2] = true;
+        shape[3, 3] = true;
 
-        var trimmed = shape.AsReadOnly().Trim(Allocator.TempJob);
-        var baseImmutable = trimmed.AsReadOnly().GetOrCreateImmutable();
-        shape.Dispose();
-        trimmed.Dispose();
+        var (_, _, trimWidth, trimHeight) = shape.GetTrimmedBound<GridShape, bool>();
+        using var trimmed = new GridShape(trimWidth, trimHeight);
+        shape.Trim<GridShape, GridShape, bool>(trimmed);
+        var baseImmutable = trimmed.GetOrCreateImmutable();
 
         const int threadCount = 30;
         const int transformsPerThread = 200;
@@ -729,18 +547,18 @@ public class ImmutableGridShapeConcurrencyTests
                 {
                     barrier.SignalAndWait();
 
-                    var random = new Random((uint)(threadId + 1));
+                    var random = new Random(threadId + 1);
                     var current = baseImmutable;
 
                     for (int i = 0; i < transformsPerThread; i++)
                     {
-                        if (random.NextBool())
+                        if (random.Next(0, 2) == 1)
                         {
                             // Verify that 4 rotations return to the same shape
                             // Do this check occasionally to avoid too much overhead
                             if (i % 10 == 0)
                             {
-                                var rot360 = current.Rotate90(Allocator.TempJob).Rotate90(Allocator.TempJob).Rotate90(Allocator.TempJob).Rotate90(Allocator.TempJob);
+                                var rot360 = current.Rotate90().Rotate90().Rotate90().Rotate90();
                                 if (rot360.Id != current.Id)
                                 {
                                     lock (errors)
@@ -756,7 +574,7 @@ public class ImmutableGridShapeConcurrencyTests
                             // Do this check occasionally to avoid too much overhead
                             if (i % 10 == 0)
                             {
-                                var doubleFlip = current.Flip(Allocator.TempJob).Flip(Allocator.TempJob);
+                                var doubleFlip = current.Flip().Flip();
                                 if (doubleFlip.Id != current.Id)
                                 {
                                     lock (errors)
@@ -781,9 +599,66 @@ public class ImmutableGridShapeConcurrencyTests
         foreach (var thread in threads) thread.Start();
         foreach (var thread in threads) thread.Join();
 
-        Assert.IsEmpty(errors, $"Errors during transformation stress test:\n{string.Join("\n", errors)}");
+        Assert.That(errors, Is.Empty, $"Errors during transformation stress test:\n{string.Join("\n", errors)}");
+    }
+
+    [Test]
+    public void StressTest_ConcurrentShapeDeduplication()
+    {
+        // Test that identical shapes created concurrently get deduplicated properly
+        const int threadCount = 16;
+        const int shapesPerThread = 50;
+        var barrier = new Barrier(threadCount);
+        var allResults = new int[threadCount][];
+
+        var threads = new Thread[threadCount];
+        for (int t = 0; t < threadCount; t++)
+        {
+            var threadIndex = t;
+            allResults[threadIndex] = new int[shapesPerThread];
+
+            threads[t] = new Thread(() =>
+            {
+                barrier.SignalAndWait();
+
+                // Each thread creates the same sequence of shapes
+                for (int i = 0; i < shapesPerThread; i++)
+                {
+                    using var shape = new GridShape(i % 3 + 1, i % 3 + 1);
+                    // Create a deterministic pattern based on index
+                    for (int j = 0; j <= i % 5; j++)
+                    {
+                        int x = j % shape.Width;
+                        int y = j / shape.Width;
+                        if (y < shape.Height)
+                            shape[x, y] = true;
+                    }
+
+                    var (_, _, trimWidth, trimHeight) = shape.GetTrimmedBound<GridShape, bool>();
+                    using var trimmed = new GridShape(trimWidth, trimHeight);
+                    shape.Trim<GridShape, GridShape, bool>(trimmed);
+                    var immutable = trimmed.GetOrCreateImmutable();
+                    allResults[threadIndex][i] = immutable.Id;
+                }
+            });
+        }
+
+        foreach (var thread in threads) thread.Start();
+        foreach (var thread in threads) thread.Join();
+
+        // Verify all threads got the same IDs for the same shapes
+        for (int i = 0; i < shapesPerThread; i++)
+        {
+            var expectedId = allResults[0][i];
+            for (int t = 1; t < threadCount; t++)
+            {
+                Assert.That(allResults[t][i], Is.EqualTo(expectedId),
+                    $"Thread {t} shape {i} got ID {allResults[t][i]}, expected {expectedId}");
+            }
+        }
+
+        TestContext.WriteLine($"Deduplication test: {threadCount} threads created {shapesPerThread} shapes each, all properly deduplicated");
     }
 
     #endregion
 }
-#endif
