+++ Packages/com.fullmetalbagel.dope-grid/Runtime/Pool.cs
@@ -1,24 +1,24 @@
 using System;
-using System.Collections.Generic;
+using System.Collections.Concurrent;
 using System.Threading;
 
 namespace DopeGrid;
 
 internal static class ArrayPool
 {
-    private static readonly Lazy<HashSet<object>> s_using = new(() => new(128), LazyThreadSafetyMode.ExecutionAndPublication);
+    private static readonly Lazy<ConcurrentDictionary<object, byte>> s_using = new(() => new(), LazyThreadSafetyMode.ExecutionAndPublication);
 
     public static T[] Rent<T>(int minimumLength)
     {
         var obj = ArrayPool<T>.Pool.Value.Rent(minimumLength);
         Array.Clear(obj, 0, obj.Length);
-        s_using.Value.Add(obj);
+        s_using.Value.TryAdd(obj, 0);
         return obj;
     }
 
     public static void Return<T>(T[] obj)
     {
-        if (s_using.Value.Remove(obj)) ArrayPool<T>.Pool.Value.Return(obj);
+        if (s_using.Value.TryRemove(obj, out _)) ArrayPool<T>.Pool.Value.Return(obj);
         else throw new InvalidOperationException("Object was not in use");
     }
 }
